Add alt under enter and tmux prefix key
It seemed that the mod/tab for the ctrl+b combination was a bit flaky.
This adds a dedicated button that sends out the often used tmux prefix
key in one go, so we do not have to make the ctrl+b rolls manually.

An alternative would be to tweak mod/tap configurations, but this seems
a bit more robust, and sufficient thumb keys are still available on the
left side.